Delete ReviewLink, refactor Review and unit test.
